WEEk 9/mini project/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-phonenumber-field==7.0.2
sqlparse==0.4.3
